ulong => uint

+++ SCGPS/SCGPS.Data/Entities/BaseEntity.cs
@@ -8,7 +8,7 @@ namespace SCGPS.Data.Entities
 {
     public class BaseEntity
     {
-        public ulong Id { get; set; }
+        public uint Id { get; set; }
         public DateTime Created { get; set; }
         public DateTime Modified { get; set; }
     }
