+++ SCGPS/SCGPS.Data/Entities/Review.cs
@@ -8,7 +8,7 @@ namespace SCGPS.Data.Entities
 {
     public class Review : BaseEntity
     {
-        public ulong MovieId { get; set; }
+        public uint MovieId { get; set; }
         public Movie Movie { get; set; }
         public string ReviewText { get; set; }
     }
