+++ SCGPS/SCGPS.Domain/Commands/ReviewSvc/AddReviewCommand.cs
@@ -8,7 +8,7 @@ namespace SCGPS.Domain.Commands.ReviewSvc
 {
     public class AddReviewCommand : ServiceCommand
     {
-        public ulong MovieId { get; set; }
+        public uint MovieId { get; set; }
         public string ReviewText { get; set; }
     }
 }
